requirements.txt
django
djangorestframework
markdown
django-filter
mysql
mysqlclient
django-cors-headers
gunicorn
pytz
dj-rest-auth
django-admin-sortable2